# week-04/day-16-mcp-agent/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Day 16 MCP Agent",
    "description": "Local agent with a real MCP server (Streamable HTTP), a discovery CLI and a streamed browser chat.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/chats": {
      "get": {
        "summary": "List Chats",
        "operationId": "list_chats_api_chats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Chat",
        "operationId": "create_chat_api_chats_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}": {
      "patch": {
        "summary": "Rename Chat",
        "operationId": "rename_chat_api_chats__chat_id__patch",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Chat",
        "operationId": "delete_chat_api_chats__chat_id__delete",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}/messages": {
      "get": {
        "summary": "Chat Messages",
        "operationId": "chat_messages_api_chats__chat_id__messages_get",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessagesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}/clear": {
      "post": {
        "summary": "Clear Chat",
        "operationId": "clear_chat_api_chats__chat_id__clear_post",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatClearResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}/tasks": {
      "get": {
        "summary": "Chat Tasks",
        "operationId": "chat_tasks_api_chats__chat_id__tasks_get",
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatTasksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "summary": "Chat Stream",
        "operationId": "chat_stream_api_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatClearResponse": {
        "properties": {
          "cleared": {
            "type": "boolean",
            "title": "Cleared"
          },
          "messages_deleted": {
            "type": "integer",
            "title": "Messages Deleted",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "cleared"
        ],
        "title": "ChatClearResponse",
        "description": "Result of clearing one chat."
      },
      "ChatCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          }
        },
        "type": "object",
        "title": "ChatCreateRequest",
        "description": "Payload of ``POST /api/chats``; an empty title gets the default."
      },
      "ChatDeleteResponse": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          },
          "stopped_tasks": {
            "type": "integer",
            "title": "Stopped Tasks",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "ChatDeleteResponse",
        "description": "Result of deleting a chat."
      },
      "ChatInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count",
            "default": 0
          },
          "active_tasks_count": {
            "type": "integer",
            "title": "Active Tasks Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ChatInfo",
        "description": "One chat as shown in the chat list."
      },
      "ChatListResponse": {
        "properties": {
          "chats": {
            "items": {
              "$ref": "#/components/schemas/ChatInfo"
            },
            "type": "array",
            "title": "Chats"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 5
          },
          "limit_reached": {
            "type": "boolean",
            "title": "Limit Reached",
            "default": false
          }
        },
        "type": "object",
        "title": "ChatListResponse",
        "description": "The chat list plus the configured limit."
      },
      "ChatMessagesResponse": {
        "properties": {
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageInfo"
            },
            "type": "array",
            "title": "Messages"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "chat_id"
        ],
        "title": "ChatMessagesResponse",
        "description": "The full stored history of one chat."
      },
      "ChatRenameRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ChatRenameRequest",
        "description": "Payload of ``PATCH /api/chats/{chat_id}``."
      },
      "ChatRequest": {
        "properties": {
          "chat_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Chat Id"
          },
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "chat_id",
          "message"
        ],
        "title": "ChatRequest",
        "description": "One chat turn of one saved chat."
      },
      "ChatTasksResponse": {
        "properties": {
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskInfo"
            },
            "type": "array",
            "title": "Tasks"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "chat_id"
        ],
        "title": "ChatTasksResponse",
        "description": "The scheduled tasks of one chat."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageInfo": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content",
          "created_at"
        ],
        "title": "MessageInfo",
        "description": "One stored chat message."
      },
...
      "TaskInfo": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "interval_seconds": {
            "type": "integer",
            "title": "Interval Seconds"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "next_run_at": {
            "type": "number",
            "title": "Next Run At"
          },
          "last_run_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "last_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Status"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "run_count": {
            "type": "integer",
            "title": "Run Count",
            "default": 0
          },
          "last_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskRunInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "query",
          "interval_seconds",
          "status",
          "created_at",
          "next_run_at"
        ],
        "title": "TaskInfo",
        "description": "One scheduled task as shown in the task panel."
      },
      "TaskRunInfo": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "ran_at": {
            "type": "number",
            "title": "Ran At"
          },
          "result_count": {
            "type": "integer",
            "title": "Result Count",
            "default": 0
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Results"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "status",
          "ran_at"
        ],
        "title": "TaskRunInfo",
        "description": "The latest stored run of a scheduled task."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}